Create Date Splitting
A date of the form 8-MAR-85 includes the name of the month, which must be translated to a number. Create a dict suitable for decoding month names to numbers. Create a function which uses string operations to split the date into 3 items using the "-" character. Translate the month, correct the year to include all of the digits.
The function will accept a date in the "dd-MMM-yy" format and respond with a tuple of ( y , m , d ).